# npool/appuser/gw/v1/banapp/banapp.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "npool/appuser/gw/v1/banapp/banapp.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Gateway"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create/ban/app": {
      "post": {
        "summary": "Admin apis",
        "operationId": "Gateway_CreateBanApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/banappv1CreateBanAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/banappv1CreateBanAppRequest"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/delete/ban/app": {
      "post": {
        "operationId": "Gateway_DeleteBanApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/banappv1DeleteBanAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/banappv1DeleteBanAppRequest"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/update/ban/app": {
      "post": {
        "operationId": "Gateway_UpdateBanApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/banappv1UpdateBanAppResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/banappv1UpdateBanAppRequest"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    }
  },
  "definitions": {
    "banappv1CreateBanAppRequest": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/v2BanAppReq",
          "title": "Permissioned by path"
        }
      }
    },
    "banappv1CreateBanAppResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/v2BanApp"
        }
      }
    },
    "banappv1DeleteBanAppRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "title": "Permissioned by path"
        }
      }
    },
    "banappv1DeleteBanAppResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/v2BanApp"
        }
      }
    },
    "banappv1UpdateBanAppRequest": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/v2BanAppReq",
          "title": "Permissioned by path"
        }
      }
    },
    "banappv1UpdateBanAppResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/v2BanApp"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v2BanApp": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "AppID": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        }
      }
    },
    "v2BanAppReq": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "AppID": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        }
      }
    }
  }
}